moved component to folders
